src/requirements.txt
biopython==1.84
pandas==2.2.3
shiny==1.2.1
shinywidgets==0.3.4
plotly==5.24.1
openpyxl==3.1.2
pyfaidx==0.8.1.3
cyvcf2==0.31.1